Local / Embedded MongoDB switch support. UUID autogenarted for Playlist.
