+++ .gitignore
@@ -1,3 +1,5 @@
+playlistmanager.iml
+
 # Compiled class file
 *.class
 
@@ -24,6 +26,8 @@ hs_err_pid*
 
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio and Webstorm
 # Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
+.idea/
+target/
 
 # User-specific stuff:
 .idea/**/workspace.xml
