--- src/main/java/com/musibondi/MongoConfig.java
@@ -1,29 +1,0 @@
-package com.musibondi;
-
-
-import com.mongodb.MongoClient;
-import cz.jirutka.spring.embedmongo.EmbeddedMongoFactoryBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.mongodb.core.MongoTemplate;
-
-import java.io.IOException;
-
-/**
- * This config will only be included in the testing part of SpringBoot, or when running with LOCAL profile.
- * FIXME: To be included only with Spring profile LOCAL
- */
-@Configuration
-public class MongoConfig {
-  private static final String MONGO_DB_URL = "localhost";
-  private static final String MONGO_DB_NAME = "embeded_db";
-
-  @Bean
-  public MongoTemplate mongoTemplate() throws IOException {
-    EmbeddedMongoFactoryBean mongo = new EmbeddedMongoFactoryBean();
-    mongo.setBindIp(MONGO_DB_URL);
-    MongoClient mongoClient = mongo.getObject();
-    MongoTemplate mongoTemplate = new MongoTemplate(mongoClient, MONGO_DB_NAME);
-    return mongoTemplate;
-  }
-}
