+++ src/main/java/com/musibondi/domain/Playlist.java
@@ -1,17 +1,19 @@
 package com.musibondi.domain;
 
 import lombok.Data;
+import lombok.Getter;
 import lombok.NonNull;
 import org.springframework.data.annotation.Id;
 
 import java.util.Collection;
+import java.util.UUID;
 
 @Data
 public class Playlist {
 
     @Id
-    @NonNull
-    private String id;
+    @Getter
+    private String id = UUID.randomUUID().toString();
 
     @NonNull
     private String name;
