+++ src/main/java/com/musibondi/mongo/EmbeddedMongoConfig.java
@@ -1,0 +1,36 @@
+package com.musibondi.mongo;
+
+
+import com.mongodb.MongoClient;
+import cz.jirutka.spring.embedmongo.EmbeddedMongoFactoryBean;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.data.mongodb.core.MongoTemplate;
+
+import java.io.IOException;
+
+/**
+ * Configuration used for embedded Mongo.
+ */
+@Profile("local")
+@Configuration
+public class EmbeddedMongoConfig {
+
+    private static final String MONGO_DB_URL = "localhost";
+    private static final String MONGO_DB_NAME = "playlist_embeded_db";
+
+    @Value("${mongo.port}")
+    private int mongoPort;
+
+    @Bean
+    public MongoTemplate mongoTemplate() throws IOException {
+        EmbeddedMongoFactoryBean mongo = new EmbeddedMongoFactoryBean();
+        mongo.setPort(mongoPort);
+        mongo.setBindIp(MONGO_DB_URL);
+        MongoClient mongoClient = mongo.getObject();
+        MongoTemplate mongoTemplate = new MongoTemplate(mongoClient, MONGO_DB_NAME);
+        return mongoTemplate;
+    }
+}
