+++ src/main/resources/application.properties
@@ -1,1 +1,24 @@
+## Spring config
+
+spring.profiles.active=local
+
+# FIXME: Uncomment when in PROD. Otherwise an embedded config is still created.
+
+## Server properties
+
 server.contextPath=/api/v1
+server.port=8080
+
+## Mongo config
+
+# Embedded ONLY
+mongo.port=27018
+
+# Not Embedded
+#spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.mongo.embedded.EmbeddedMongoAutoConfiguration
+#spring.data.mongodb.host=localhost
+#spring.data.mongodb.port=27017
+#spring.data.mongodb.database=playlist_db
+#spring.data.mongodb.username=user
+#spring.data.mongodb.password=pass
+
