+++ src/main/resources/banner.txt
@@ -1,0 +1,42 @@
+       :yho.                    /-                                                                                                                    
+      oMMMMMh-               y/ -d.                      ``..----..``                                                                                 
+     :MMMMMMMMy.             .m- :d`                     oMNNNNmmmdddhy-                                                                              
+     sMMMNs-.-oN/             -d.:ddmo                    ss/::-------::.                                                                             
+     yMMM:     /Ms           `smMMMMMh.       `.-:///::::::s:--....``````.                                                                            
+     sMMd       NMs           +MNN+-do:///////:-.`          +`         ``--.                                                                          
+     -MMh       NMM-       `.:/o+dh/:d/.`                    +`           .-..`                                                                       
+      hMm      :MMMs`-/++++/-.   `d+/hMMm`                   `o`           .. ``                                                                      
+      `mM:   `-dMMMm/-`          omMMMMMh:       -/      `.--::s--........odh-  `                                                                     
+       -Nmo++/mMMMMy             -NNMo.+h`    -h.:mo/////:-.`  `+         +NNm:-`.                                                                    
+ .:++++:-dm-.hMMMMM/              ` +d``yy/+++/om-.h:         -/+s         -+/  `-+` `                                                                
++/.      `yNNMMMMMm`            `-:/+dd+:y/     oy.+Nmh`    .dMMNNo               :o- `                                                               
+          +MMMMMMM:       .:++++/-`   h:       `sNMMMMm:    sMMMMNh  `..--------.`.:s/`                                                               
+        `yMMMMMMM/ `-/++++:.                    oMNm/:d`   `:dNmhs::-..`       ``-/-++/``                                                             
+       `dMMMMMMMd++/-`                           .`d/-yd+s+:.`                   `. --++s/`                                                           
+      .mMMMMMMmyN/                           `-/++++m+/mNM+                      .  - :o:`.                                                           
+  ./+sNMMMMMMy` +Ns`                   .:++++/-`   :dMMMMMy`                   `./--/``/-.``                             ````-.``````                 
+++:.`mMMMMMM+    :Nd``-/ooso+:.  .:++++:.          .NMMs-so           `-:::::::--:`:../:ho.--                  ``````````    .y:```  ``.soo:          
+    hMMMMMN/      :NNMMMMMMMMMMNNs.                 .-/h` y+   `-////:-.        .` -  /:MM/:-           ````.//++++```````````++:```-```-``-  `       
+   /MMMMMN:     .yMMMMMMMMMMMMMMMMd:                   +y`-mo++/-`              . -- --`/y/ . .``  `....`  `:/:..`.`  ```````-+.-.`...`oy``..``       
+   mMMMMM+     -mMMMMMMNo///+ohNMMMMs               ./++hd-`                    .`:-.:`.:/`   ``:..`.`....` `-`````:``     `.yy...`.//`````s/`:`      
+  .MMMMMd`./++oMMMMMh-:Nd.     `/dMMMo        `:++++:.   `                     `-+soo///+-::.```+.`..- `:.`/d+   /d+```````     ```.s/` ````..-``     
+  -MMMMMho/.  /MMMMo   .dN:       oMMM.  ./+o+/-`                         .////:...:...o`  ::``/+ `.oo.``.`...```    ````````````     ````````. /-    
+:+sNMMMM-     +MMMd     `yMo       hMMyo+:.                         `-/++/-`      `-:++.:`-:.`````` :..-y/`  `.``````                      ```` :`    
+-` yMMMM-     `NMMy       +Ny`./+oohMM.                         -/+++-`            `-/:..:--+`  ``..`  -+...`                                 .`-.    
+   `mMMMs      .mMN:    `-/hNMo.   hMs                     ./oo+:`                  .`.--`.````   `...`                                  .   `:`-     
+    .dMMN-       /dMyoso+:. .dN/  +My                 ./+++:.                       `.:-.`  ``````                                      `. --`...     
+     `sMMN:   .:+oo+/osys:   `yMoyN/             `:ooo/.                             `.....``                                    `      `-..-.- ``    
+       -hMMhyo/-`            `/NMN-         `:ooo/.                                                                             `-`.``-:-``:`.`       
+    -/ooo/odMms/-`      .-/sdNy//Nm-   `:+oo+-`                                                                               `.-``.```..````         
+/ooo/.      `/ohdNNNNNNNmhs+-    .dNsoo+-`                                                                                    ..--`````````           
+.                 `````      `-+oo+sMs`                                                                                      `..-..``                 
+                        `-+ooo:`    /Nd.                                                                                      ```....`.:.             
+                   `-+ooo:.          .mm.                                                                                         ``.-:`:`` :.   `.   
+              `-+ooo/.  `-//:.        -Md                                                                                            ``````.:``.`:.-: 
+         `-/oso/.     .sNMMMMMd:       hM-                                                                                                  ``--`.`.: 
+    `-+oso/.         .NMMMMMMMMM:      hM:                                                                                                          ``
+-+ooo/.              hMMMMMMMMMMo     -Nm`                                                                                                            
+/.                   dMMMMMMMMMd`    :Nm.                                                                                                             
+                     /MMMMMMMd+`  `/hNs`                                                                                                              
+                      :dMMMMd+/+ohNd+`                                                                                                                
+                        .+shddhyo/.
